+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CentOS-6 6.8 x86_64 - Memcached 1.4.
 - Adds a change log (`CHANGELOG.md`).
 - Adds support for semantic version numbered tags.
 - Adds minor code style changes to the Makefile for readability.
+- Adds support for running `shpec` functional tests with `make test`.
 
 ### 1.0.0 - 2016-11-23